Move evm fees (evmSummary.tsx) into its own file
