+++ webapp/app/[locale]/tunnel/_components/deposit.tsx
@@ -39,7 +39,8 @@ import { canSubmit, getTotal } from '../_utils'
 import { BtcFees } from './btcFees'
 import { ConnectEvmWallet } from './connectEvmWallet'
 import { Erc20TokenApproval } from './erc20TokenApproval'
-import { EvmSummary, FormContent, TunnelForm } from './form'
+import { EvmSummary } from './evmSummary'
+import { FormContent, TunnelForm } from './form'
 import { ReceivingAddress } from './receivingAddress'
 import { SubmitWithTwoWallets } from './submitWithTwoWallets'
